refactor(blorktools): rig factory
