+++ packages/blorktools/src/asset_debugger/ui/rig-panel.js
@@ -6,22 +6,37 @@
  */
 import * as THREE from 'three';
 import { getState } from '../core/state.js';
-
-// Global variables
-let rigDetails = null;
-let bones = [];
-let boneVisualsGroup = null;
-let furthestBoneHandle = null;
-let raycaster = new THREE.Raycaster();
-let mouse = new THREE.Vector2();
-let hoveredHandle = null;
+import { 
+    analyzeGltfModel, 
+    deduplicateItems, 
+    findAssociatedBone, 
+    createRigVisualization, 
+    createBoneMesh, 
+    createBoneUpdateFunction, 
+    findFarthestBone, 
+    addControlHandleToFurthestBone,
+    bones,
+    boneVisualsGroup,
+    boneMaterial,
+    boneSideMaterial,
+    furthestBoneHandle,
+    rigDetails,
+    isDragging,
+    rigOptions,
+    updateRigDetails,
+    setIsDragging
+} from '../core/rig/rig-factory.js';
 
 // Reusable objects for position and rotation operations
 let worldPos = new THREE.Vector3();
 let worldRot = new THREE.Quaternion();
 
+// Raycaster for mouse interaction
+let raycaster = new THREE.Raycaster();
+let mouse = new THREE.Vector2();
+let hoveredHandle = null;
+
 // Drag state tracking
-let isDragging = false;
 let dragStartPosition = new THREE.Vector3();
 let dragPlane = new THREE.Plane();
 let dragOffset = new THREE.Vector3();
@@ -32,29 +47,6 @@ let dragTargetPosition = new THREE.Vector3();
 let selectedBones = new Set();
 let lockedBones = new Map(); // Maps bone.uuid to {bone, originalRotation}
 let labelGroup = null;
-
-// Material colors
-const normalColor = 0xff0000; // Red
-const hoverColor = 0x00ff00;  // Green
-const activeColor = 0x0000ff; // Blue - for when dragging
-
-// Rig options
-const rigOptions = {
-    displayRig: false, // Default to not visible
-    forceZ: false,
-    wireframe: true,
-    primaryColor: 0xFF00FF, // Magenta
-    secondaryColor: 0xFFFF00, // Yellow
-    jointColor: 0x00FFFF, // Cyan
-    showJointLabels: false, // Default to hidden
-    normalColor: 0xFF0000, // Red - for control handles normal state
-    hoverColor: 0x00FF00,  // Green - for control handles hover state
-    activeColor: 0x0000FF  // Blue - for control handles active/dragging state
-};
-
-// Material references
-let boneMaterial = null;
-let boneSideMaterial = null;
 
 // IK settings
 const IK_CHAIN_LENGTH = 3; // Maximum bones in IK chain
@@ -75,653 +67,6 @@ let jointSettingsDebug = true;
 // Add global state tracking for joint settings
 let allJointsInPreviousState = true;
 let jointPreviousValues = new Map(); // Map of joint name to previous value
-
-/**
- * Analyze the rig data in a GLTF model
- * @param {Object} gltf - The loaded GLTF model data
- * @returns {Object} Analyzed rig details
- */
-function analyzeGltfModel(gltf) {
-    console.log('analyzeGltfModel called with:', gltf);
-    
-    if (!gltf || !gltf.scene) {
-        console.error('Invalid GLTF model:', gltf);
-        return null;
-    }
-    
-    console.log('Analyzing GLTF model for rig information...');
-    console.log('GLTF scene structure:', gltf.scene);
-    
-    const rawDetails = {
-        bones: [],
-        rigs: [],
-        roots: [],
-        controls: [], // Handles/Controls
-        joints: []    // Add joints array to store joint data
-    };
-    
-    // Extract scene information
-    const scene = gltf.scene;
-    
-    // Helper function to traverse the scene
-    const traverseNode = (node, parentType = null) => {
-        console.log('Traversing node:', node.name, 'type:', node.type, 'isBone:', node.isBone);
-        
-        // Check if the node is a bone
-        if (node.isBone || node.name.toLowerCase().includes('bone')) {
-            console.log('Found bone:', node.name);
-            rawDetails.bones.push({
-                name: node.name,
-                position: node.position ? [node.position.x, node.position.y, node.position.z] : null,
-                rotation: node.rotation ? [node.rotation.x, node.rotation.y, node.rotation.z] : null,
-                parentName: parentType === 'bone' ? node.parent.name : null
-            });
-            parentType = 'bone';
-        }
-        
-        // Check if the node is a rig
-        if (node.name.toLowerCase().includes('rig') || node.name.toLowerCase().includes('armature')) {
-            console.log('Found rig:', node.name);
-            rawDetails.rigs.push({
-                name: node.name,
-                position: node.position ? [node.position.x, node.position.y, node.position.z] : null,
-                childCount: node.children ? node.children.length : 0
-            });
-            parentType = 'rig';
-        }
-        
-        // Check if the node is a root
-        if (node.name.toLowerCase().includes('root')) {
-            console.log('Found root:', node.name);
-            rawDetails.roots.push({
-                name: node.name,
-                position: node.position ? [node.position.x, node.position.y, node.position.z] : null
-            });
-            parentType = 'root';
-        }
-        
-        // Check if the node is a control/handle
-        if (node.name.toLowerCase().includes('control') || 
-            node.name.toLowerCase().includes('ctrl') || 
-            node.name.toLowerCase().includes('handle')) {
-            console.log('Found control:', node.name);
-            rawDetails.controls.push({
-                name: node.name,
-                position: node.position ? [node.position.x, node.position.y, node.position.z] : null,
-                type: node.name.toLowerCase().includes('control') ? 'control' : 
-                      node.name.toLowerCase().includes('ctrl') ? 'ctrl' : 'handle'
-            });
-            parentType = 'control';
-        }
-        
-        // Traverse children
-        if (node.children) {
-            node.children.forEach(child => traverseNode(child, parentType));
-        }
-    };
-    
-    // Start traversal from the scene
-    console.log('Starting scene traversal');
-    scene.traverse(node => traverseNode(node));
-    
-    // Process raw details to deduplicate items
-    const result = {
-        bones: deduplicateItems(rawDetails.bones),
-        rigs: deduplicateItems(rawDetails.rigs),
-        roots: deduplicateItems(rawDetails.roots),
-        controls: deduplicateItems(rawDetails.controls),
-        joints: deduplicateItems(rawDetails.joints)
-    };
-    
-    console.log('Rig analysis results:', result);
-    return result;
-}
-
-/**
- * Deduplicate items by name, counting duplicates
- * @param {Array} items - Array of items to deduplicate
- * @returns {Array} Deduplicated items with count property
- */
-function deduplicateItems(items) {
-    const uniqueItems = new Map();
-    
-    items.forEach(item => {
-        const key = item.name;
-        if (uniqueItems.has(key)) {
-            const existingItem = uniqueItems.get(key);
-            existingItem.count = (existingItem.count || 1) + 1;
-        } else {
-            uniqueItems.set(key, {...item, count: 1});
-        }
-    });
-    
-    return Array.from(uniqueItems.values());
-}
-
-/**
- * Find associated bone for a control by its name
- * @param {String} controlName - Name of the control
- * @param {Array} bones - Array of bones to search
- * @returns {Object|null} Associated bone or null if not found
- */
-function findAssociatedBone(controlName, bones) {
-    // Try matching by name
-    const boneName = controlName.replace('control', 'bone')
-                                .replace('ctrl', 'bone')
-                                .replace('handle', 'bone');
-    
-    let matchedBone = null;
-    bones.forEach(bone => {
-        if (bone.name === boneName || bone.name.includes(boneName) || boneName.includes(bone.name)) {
-            matchedBone = bone;
-        }
-    });
-    
-    return matchedBone;
-}
-
-/**
- * Create a visualization of bones in the 3D scene
- * @param {Object} model - The 3D model to visualize
- * @param {Object} scene - The Three.js scene
- */
-function createRigVisualization(model, scene) {
-    if (!model || !scene) return;
-    
-    console.log('Creating rig visualization...');
-    
-    // Clear any existing rig visualization
-    clearRigVisualization(scene);
-    
-    // Reset arrays
-    bones = [];
-    
-    // Find all bones
-    let armature = null;
-    
-    // Find armature first
-    model.traverse(node => {
-        if ((node.name.toLowerCase().includes('rig') || 
-             node.name.toLowerCase().includes('armature')) && !armature) {
-            armature = node;
-            console.log('Found armature:', node.name);
-        }
-    });
-    
-    // First pass: collect all bones from armature if found
-    if (armature) {
-        armature.traverse(node => {
-            if (node.isBone || node.name.toLowerCase().includes('bone')) {
-                // Store initial rotation for reset functionality
-                node.userData.initialRotation = {
-                    x: node.rotation.x,
-                    y: node.rotation.y,
-                    z: node.rotation.z,
-                    order: node.rotation.order
-                };
-                // Log whether this is a root bone
-                if (node.name.toLowerCase().includes('root')) {
-                    console.log('Found ROOT bone:', node.name);
-                } else {
-                    console.log('Found bone:', node.name);
-                }
-                bones.push(node);
-            }
-        });
-    }
-    
-    // If no bones were found in the armature, search the entire model
-    if (bones.length === 0) {
-        console.log('No bones found in armature, searching entire model');
-        model.traverse(node => {
-            if (node.isBone || node.name.toLowerCase().includes('bone')) {
-                // Store initial rotation for reset functionality
-                node.userData.initialRotation = {
-                    x: node.rotation.x,
-                    y: node.rotation.y,
-                    z: node.rotation.z,
-                    order: node.rotation.order
-                };
-                // Log whether this is a root bone
-                if (node.name.toLowerCase().includes('root')) {
-                    console.log('Found ROOT bone in model:', node.name);
-                } else {
-                    console.log('Found bone in model:', node.name);
-                }
-                bones.push(node);
-            }
-        });
-    }
-    
-    // If still no bones, exit
-    if (bones.length === 0) {
-        console.log('No bones found in model');
-        return;
-    }
-    
-    // Create materials for bone visualization
-    boneMaterial = new THREE.MeshPhongMaterial({
-        color: rigOptions.primaryColor,
-        emissive: 0x072534,
-        side: THREE.DoubleSide,
-        flatShading: true,
-        wireframe: rigOptions.wireframe,
-        transparent: true,
-        opacity: 0.8  // Increased opacity for better visibility
-    });
-    
-    boneSideMaterial = new THREE.MeshPhongMaterial({
-        color: rigOptions.secondaryColor,
-        emissive: 0x072534,
-        side: THREE.DoubleSide,
-        flatShading: true,
-        wireframe: rigOptions.wireframe,
-        transparent: true,
-        opacity: 0.8  // Increased opacity for better visibility
-    });
-    
-    // Joint material (now separate from primary/secondary colors)
-    const jointMaterial = new THREE.MeshPhongMaterial({
-        color: rigOptions.jointColor,
-        emissive: 0x072534,
-        side: THREE.DoubleSide,
-        flatShading: true,
-        wireframe: rigOptions.wireframe,
-        transparent: true,
-        opacity: 0.8  // Slightly more opaque for better visibility
-    });
-    
-    // Create a group to hold all bone visualizations
-    boneVisualsGroup = new THREE.Group();
-    boneVisualsGroup.name = "BoneVisualizations";
-    boneVisualsGroup.visible = rigOptions.displayRig;
-    scene.add(boneVisualsGroup);
-    
-    // Create axis indicator
-    const state = getState();
-    if (state.renderer && state.camera) {
-        createAxisIndicator(scene, state.camera, state.renderer);
-    }
-    
-    // Group bones by parent for easier bone pair creation
-    const bonesByParent = new Map();
-    
-    // Filter out control bones that we don't want to visualize
-    const visualizableBones = bones.filter(bone => 
-        !(bone.name.toLowerCase().includes('control') || 
-          bone.name.toLowerCase().includes('ctrl') || 
-          bone.name.toLowerCase().includes('handle'))
-    );
-    
-    // Identify root bones
-    const rootBones = visualizableBones.filter(bone => 
-        bone.name.toLowerCase().includes('root')
-    );
-    
-    // Group bones by parent for easier bone pair creation
-    visualizableBones.forEach(bone => {
-        if (bone.parent) {
-            const parentId = bone.parent.uuid;
-            if (!bonesByParent.has(parentId)) {
-                bonesByParent.set(parentId, []);
-            }
-            bonesByParent.get(parentId).push(bone);
-        }
-    });
-    
-    // Calculate model scale for appropriate bone visualization size
-    const bbox = new THREE.Box3().setFromObject(model);
-    const size = new THREE.Vector3();
-    bbox.getSize(size);
-    const modelScale = size.length() * 0.02;
-    const boneRadius = Math.max(0.02, modelScale * 0.3);
-    
-    // Create visual bone meshes between parent-child bone pairs
-    visualizableBones.forEach(bone => {
-        // Skip if this bone is not in our scene
-        if (!scene.getObjectById(bone.id)) return;
-        
-        // Get current bone position
-        const bonePos = new THREE.Vector3();
-        bone.getWorldPosition(bonePos);
-        
-        // Check if this bone has children in our bone list
-        const childBones = bonesByParent.get(bone.uuid) || [];
-        
-        // If this bone has child bones, create a visual bone for each connection
-        childBones.forEach(childBone => {
-            // Skip control bones
-            if (childBone.name.toLowerCase().includes('control') || 
-                childBone.name.toLowerCase().includes('ctrl') || 
-                childBone.name.toLowerCase().includes('handle')) {
-                return;
-            }
-            
-            // Get child bone position
-            const childPos = new THREE.Vector3();
-            childBone.getWorldPosition(childPos);
-            
-            // Calculate distance and direction
-            const distance = bonePos.distanceTo(childPos);
-            
-            // Only create visual bone if distance is not zero
-            if (distance > 0.001) {
-                // Create a group for the bone mesh
-                const boneGroup = new THREE.Group();
-                boneVisualsGroup.add(boneGroup);
-                
-                // Position bone group at parent bone position
-                boneGroup.position.copy(bonePos);
-                
-                // Make the bone look at the child
-                const direction = new THREE.Vector3().subVectors(childPos, bonePos);
-                boneGroup.lookAt(childPos);
-                
-                // Rotate to align with standard Three.js cylinder orientation
-                boneGroup.rotateX(Math.PI/2);
-                
-                // Pass both materials (primary and secondary) for alternating sides
-                createBoneMesh(boneGroup, boneRadius, boneRadius, distance, jointMaterial, boneMaterial, boneSideMaterial);
-                
-                // Store reference to the bone connection
-                boneGroup.userData.parentBone = bone;
-                boneGroup.userData.childBone = childBone;
-                
-                // Add update function
-                boneGroup.userData.isVisualBone = true;
-                boneGroup.userData.updatePosition = createBoneUpdateFunction(boneGroup);
-                
-                // Store the joint data for the rig details panel
-                if (rigDetails && rigDetails.joints) {
-                    rigDetails.joints.push({
-                        name: `Joint_${bone.name}_to_${childBone.name}`,
-                        parentBone: bone.name,
-                        childBone: childBone.name,
-                        position: [bonePos.x, bonePos.y, bonePos.z],
-                        count: 1
-                    });
-                }
-            }
-        });
-    });
-    
-    // Create root bone visualization (as a standalone puck)
-    rootBones.forEach(rootBone => {
-        // Get root bone position
-        const rootPos = new THREE.Vector3();
-        rootBone.getWorldPosition(rootPos);
-        
-        console.log(`Creating standalone root visualization for: ${rootBone.name} at position:`, rootPos);
-        
-        // Create a group for the root visualization
-        const rootGroup = new THREE.Group();
-        rootGroup.position.copy(rootPos);
-        boneVisualsGroup.add(rootGroup);
-        
-        // Create a standalone puck for the root
-        const rootPuckSize = boneRadius * 2.5;
-        // Use a disk-like geometry (short, wide cylinder)
-        const puckGeometry = new THREE.CylinderGeometry(rootPuckSize, rootPuckSize, rootPuckSize * 0.2, 32);
-        const puckMaterial = jointMaterial.clone();
-        
-        const rootPuck = new THREE.Mesh(puckGeometry, puckMaterial);
-        // No rotation needed - by default the cylinder is already oriented with Y-axis up
-        // which means the flat sides will be parallel to the ground
-        rootPuck.userData.isRootJoint = true;
-        rootPuck.userData.bonePart = 'cap';
-        rootPuck.userData.isJoint = true;
-        rootPuck.userData.boneName = rootBone.name;
-        
-        // Make root puck render on top
-        rootPuck.renderOrder = 25; // Higher than normal joints (10-15) but lower than handle (30)
-        
-        // Add to root group
-        rootGroup.add(rootPuck);
-        
-        // Store reference to the bone
-        rootGroup.userData.rootBone = rootBone;
-        
-        // Add update function
-        rootGroup.userData.isVisualBone = true;
-        rootGroup.userData.updatePosition = () => {
-            if (rootGroup.userData.rootBone) {
-                const pos = new THREE.Vector3();
-                rootGroup.userData.rootBone.getWorldPosition(pos);
-                rootGroup.position.copy(pos);
-            }
-        };
-        
-        // Store the root joint data for the rig details panel
-        if (rigDetails && rigDetails.joints) {
-            rigDetails.joints.push({
-                name: `Root_Joint_${rootBone.name}`,
-                parentBone: "Scene Root",
-                childBone: rootBone.name,
-                position: [rootPos.x, rootPos.y, rootPos.z],
-                count: 1,
-                isRoot: true
-            });
-        }
-        
-        console.log(`Root visualization created for ${rootBone.name}`);
-    });
-    
-    // Find the furthest bone from the root and add a control handle
-    const furthestBone = findFarthestBone();
-    if (furthestBone) {
-        addControlHandleToFurthestBone(furthestBone, scene, modelScale);
-    }
-    
-    // Create labels for all joints
-    if (rigOptions.showJointLabels) {
-        console.log('Setting up joint labels');
-        createJointLabels(scene);
-    }
-    
-    // Set up mouse event listeners for hover effect
-    setupMouseListeners(scene);
-    
-    console.log('Rig visualization created with', bones.length, 'bones');
-}
-
-/**
- * Create bone mesh with joints
- * @param {Object} parent - Parent THREE.Group to add the bone to
- * @param {Number} radiusTop - Top radius of the bone
- * @param {Number} radiusBottom - Bottom radius of the bone
- * @param {Number} height - Height of the bone
- * @param {Material} capMaterial - Material for bone caps
- * @param {Material} sideMaterial - Material for bone sides
- * @param {Material} alternateSideMaterial - Material for alternate sides
- */
-function createBoneMesh(parent, radiusTop, radiusBottom, height, capMaterial, sideMaterial, alternateSideMaterial) {
-    // First create a cylinder with 8 segments
-    const cylinderGeometry = new THREE.CylinderGeometry(radiusTop, radiusBottom, height, 8, 1, false);
-    
-    // Create all sides in one group
-    const sidesGroup = new THREE.Group();
-    sidesGroup.position.y = height / 2;
-    parent.add(sidesGroup);
-    
-    // Split the cylinder into 8 segments for alternating colors
-    for (let i = 0; i < 8; i++) {
-        // Create a segment
-        const segmentGeometry = new THREE.CylinderGeometry(
-            radiusTop, radiusBottom, height, 1, 1, false,
-            (Math.PI * 2 * i) / 8,
-            Math.PI * 2 / 8
-        );
-        
-        // Use alternating materials based on segment index
-        const material = (i % 2 === 0) ? sideMaterial.clone() : alternateSideMaterial.clone();
-        
-        // Create the segment mesh
-        const segment = new THREE.Mesh(segmentGeometry, material);
-        segment.userData.bonePart = 'side';
-        segment.userData.sideType = (i % 2 === 0) ? 'primary' : 'secondary';
-        segment.userData.segmentIndex = i;
-        
-        sidesGroup.add(segment);
-    }
-    
-    // Check if this is a root bone - either through parent/child data or name
-    let isRootBone = false;
-    
-    // Check for root in parent data
-    if (parent.userData.childBone && parent.userData.childBone.name.toLowerCase().includes('root')) {
-        isRootBone = true;
-    }
-    
-    // Check for root in parent data (parent bone)
-    if (parent.userData.parentBone && parent.userData.parentBone.name.toLowerCase().includes('root')) {
-        isRootBone = true;
-    }
-    
-    // Create top joint sphere or puck
-    let topSphere;
-    
-    if (isRootBone) {
-        console.log("Creating ROOT PUCK for joint at top of bone connection");
-        // Create a puck for root joints - shorter and wider for a flatter appearance
-        const puckGeometry = new THREE.CylinderGeometry(radiusTop * 2.2, radiusTop * 2.2, radiusTop * 0.2, 32);
-        topSphere = new THREE.Mesh(puckGeometry, capMaterial.clone());
-        // No rotation - default cylinder orientation has the flat sides parallel to the ground
-    } else {
-        // Normal sphere for regular joints
-        const sphereGeometry = new THREE.SphereGeometry(radiusTop * 1.2, 8, 8);
-        topSphere = new THREE.Mesh(sphereGeometry, capMaterial.clone());
-    }
-    
-    topSphere.position.y = height;
-    topSphere.userData.bonePart = 'cap';
-    topSphere.userData.isJoint = true;
-    topSphere.userData.isRootJoint = isRootBone;
-    
-    // Ensure material settings
-    topSphere.material.wireframe = capMaterial.wireframe;
-    topSphere.material.transparent = true;
-    topSphere.material.opacity = 0.9;
-    topSphere.material.color.setHex(rigOptions.jointColor);
-    
-    // Make sure joint spheres render on top of bones
-    topSphere.renderOrder = isRootBone ? 15 : 10; // Higher render order for root pucks
-    parent.add(topSphere);
-    
-    // Store child bone name for label reference
-    if (parent.userData.childBone) {
-        topSphere.userData.boneName = parent.userData.childBone.name;
-    }
-    
-    // Create bottom joint - sphere or puck based on whether it's a root
-    let bottomSphere;
-    
-    if (isRootBone) {
-        console.log("Creating ROOT PUCK for joint at bottom of bone connection");
-        // Create a puck for root joints - flatter for better horizontal appearance
-        const puckGeometry = new THREE.CylinderGeometry(radiusBottom * 2.2, radiusBottom * 2.2, radiusBottom * 0.2, 32);
-        bottomSphere = new THREE.Mesh(puckGeometry, capMaterial.clone());
-        // No rotation - cylinder's default orientation has flat sides parallel to ground
-    } else {
-        // Normal sphere for regular joints
-        const sphereGeometry = new THREE.SphereGeometry(radiusBottom * 1.2, 8, 8);
-        bottomSphere = new THREE.Mesh(sphereGeometry, capMaterial.clone());
-    }
-    
-    bottomSphere.userData.bonePart = 'cap';
-    bottomSphere.userData.isJoint = true;
-    bottomSphere.userData.isRootJoint = isRootBone;
-    
-    // Ensure material settings 
-    bottomSphere.material.wireframe = capMaterial.wireframe;
-    bottomSphere.material.transparent = true;
-    bottomSphere.material.opacity = 0.9;
-    bottomSphere.material.color.setHex(rigOptions.jointColor);
-    
-    // Make sure joint spheres render on top of bones
-    bottomSphere.renderOrder = isRootBone ? 15 : 10; // Higher render order for root pucks
-    parent.add(bottomSphere);
-    
-    // Store parent bone name for label reference
-    if (parent.userData.parentBone) {
-        bottomSphere.userData.boneName = parent.userData.parentBone.name;
-    }
-}
-
-/**
- * Create a function to update bone visuals based on bone movement
- * @param {Object} boneGroup - The visual bone group to update
- * @returns {Function} Update function for the bone
- */
-function createBoneUpdateFunction(boneGroup) {
-    return () => {
-        if (boneGroup.userData.parentBone && boneGroup.userData.childBone) {
-            const parentPos = new THREE.Vector3();
-            const childPos = new THREE.Vector3();
-            
-            boneGroup.userData.parentBone.getWorldPosition(parentPos);
-            boneGroup.userData.childBone.getWorldPosition(childPos);
-            
-            // Update position and orientation
-            boneGroup.position.copy(parentPos);
-            
-            // Make the bone look at the child
-            const direction = new THREE.Vector3().subVectors(childPos, parentPos);
-            if (direction.lengthSq() > 0.001) {
-                boneGroup.lookAt(childPos);
-                boneGroup.rotateX(Math.PI/2);
-                
-                // Update scale to match new length
-                const distance = parentPos.distanceTo(childPos);
-                const children = boneGroup.children;
-                for (let i = 0; i < children.length; i++) {
-                    if (children[i].userData.bonePart === 'side') {
-                        children[i].scale.set(1, distance / children[i].geometry.parameters.height, 1);
-                        children[i].position.y = distance / 2;
-                    } else if (children[i].userData.bonePart === 'cap' && children[i].position.y > 0) {
-                        children[i].position.y = distance;
-                    }
-                }
-            }
-        }
-    };
-}
-
-/**
- * Find the furthest bone from the root
- * @returns {Object} The furthest bone
- */
-function findFarthestBone() {
-    if (!bones.length) return null;
-    
-    // Find bones with no children (end effectors)
-    const endBones = [];
-    
-    bones.forEach(bone => {
-        let isEndBone = true;
-        // Check if this bone has any child bones
-        for (let i = 0; i < bone.children.length; i++) {
-            const child = bone.children[i];
-            if (child.isBone || child.name.toLowerCase().includes('bone')) {
-                isEndBone = false;
-                break;
-            }
-        }
-        
-        if (isEndBone) {
-            endBones.push(bone);
-        }
-    });
-    
-    // If we found end bones, return the first one
-    if (endBones.length > 0) {
-        console.log('Found end bone:', endBones[0].name);
-        return endBones[0];
-    }
-    
-    // If we couldn't identify end bones, just return the last bone in the array
-    console.log('No end bones found, using last bone:', bones[bones.length - 1].name);
-    return bones[bones.length - 1];
-}
 
 /**
  * Set up mouse listeners for handle interaction
@@ -801,7 +146,7 @@ function setupMouseListeners(scene) {
 function startDrag(intersection, handle) {
     if (!handle) return;
     
-    isDragging = true;
+    setIsDragging(true);
     dragTarget = handle;
     
     // Update material to active color
@@ -875,7 +220,7 @@ function handleDrag() {
 function stopDrag() {
     if (!isDragging || !dragTarget) return;
     
-    isDragging = false;
+    setIsDragging(false);
     
     // Reset material to normal or hover color based on current state
     if (dragTarget.material) {
@@ -1758,7 +1103,7 @@ function updateRigPanel() {
     // Clear any existing analysis if we're explicitly updating
     if (rigDetails) {
         console.log('Clearing existing rig details for fresh analysis');
-        rigDetails = null;
+        updateRigDetails(null);
     }
     
     // If we don't have rig details yet, try to analyze the model
@@ -1770,9 +1115,12 @@ function updateRigPanel() {
             const gltfData = { scene: state.model };
             console.log('Created GLTF-like object for analysis:', gltfData);
             
-            // Analyze the model to extract rig information
-            rigDetails = analyzeGltfModel(gltfData);
-            console.log('Rig analysis complete, results:', rigDetails);
+            // Analyze the model to extract rig information using the imported function
+            const newRigDetails = analyzeGltfModel(gltfData);
+            console.log('Rig analysis complete, results:', newRigDetails);
+            
+            // Update the rig details using the exported function
+            updateRigDetails(newRigDetails);
             
             // Create the rig visualization if we have bones
             if (rigDetails && rigDetails.bones && rigDetails.bones.length > 0) {
@@ -1806,45 +1154,6 @@ function updateRigPanel() {
     
     // Create the rig details content
     createRigDetailsContent(rigContent, rigDetails);
-}
-
-/**
- * Add a control handle to the furthest bone
- * @param {Object} bone - The bone to add the handle to
- * @param {Object} scene - The Three.js scene
- * @param {Number} modelScale - Scale factor for the handle size
- */
-function addControlHandleToFurthestBone(bone, scene, modelScale) {
-    const handleSize = modelScale * 2.6; // Double the size (from 1.3 to 2.6)
-    const geometry = new THREE.SphereGeometry(handleSize, 16, 16);
-    const material = new THREE.MeshPhongMaterial({
-        color: rigOptions.normalColor, // Use color from rigOptions
-        transparent: true,
-        opacity: 0.7,
-        wireframe: false
-    });
-    
-    furthestBoneHandle = new THREE.Mesh(geometry, material);
-    furthestBoneHandle.name = "FurthestBoneHandle";
-    scene.add(furthestBoneHandle);
-    
-    // Position at the furthest bone
-    const bonePos = new THREE.Vector3();
-    bone.getWorldPosition(bonePos);
-    furthestBoneHandle.position.copy(bonePos);
-    
-    // Add information about which bone it controls
-    furthestBoneHandle.userData.controlledBone = bone;
-    furthestBoneHandle.userData.isControlHandle = true;
-    furthestBoneHandle.userData.updatePosition = () => {
-        if (furthestBoneHandle.userData.controlledBone && !isDragging) {
-            const controlledBonePos = new THREE.Vector3();
-            furthestBoneHandle.userData.controlledBone.getWorldPosition(controlledBonePos);
-            furthestBoneHandle.position.copy(controlledBonePos);
-        }
-    };
-    
-    console.log('Added control handle to furthest bone:', bone.name);
 }
 
 /**
@@ -3183,86 +2492,11 @@ function updateJointSettingsState() {
  * @param {HTMLElement} button - The button to enable
  */
 function enableApplyButton(button) {
-    button.disabled = false;
-    button.style.backgroundColor = 'rgba(66, 133, 244, 0.8)';
-    button.style.color = 'white';
-    button.style.cursor = 'pointer';
-    button.style.opacity = '1';
+    if (button) {
+        button.removeAttribute('disabled');
+        button.classList.remove('disabled');
+    }
 }
-
-// Export functions
-export { analyzeGltfModel, updateRigAnimation, updateAllBoneMatrices, updateRigPanel };
-
-// Global event listener for axis indicator mode changes
-document.addEventListener('axisIndicatorModeChange', function(e) {
-    const mode = e.detail.mode;
-    const intensity = e.detail.intensity !== undefined ? e.detail.intensity : 0.7;
-    console.log('Axis indicator mode change event received:', mode, 'intensity:', intensity);
-    
-    // Get current scene, camera, and renderer from state
-    const state = getState();
-    const scene = state.scene;
-    const camera = state.camera;
-    const renderer = state.renderer;
-    
-    if (!scene || !camera || !renderer) {
-        console.error('Cannot change axis indicator mode: scene, camera or renderer not available', {
-            scene: !!scene,
-            camera: !!camera,
-            renderer: !!renderer,
-        });
-        return;
-    }
-    
-    // Apply the mode change
-    if (mode === 'windowed') {
-        // Hide embedded version
-        if (state.embeddedAxisIndicator) {
-            state.embeddedAxisIndicator.active = false;
-        }
-        
-        // Show windowed version if it exists (or create it)
-        if (!state.axisScene) {
-            createAxisIndicator(scene, camera, renderer);
-        }
-        
-        // Show the windowed version
-        const axisContainer = document.getElementById('axis-indicator-container');
-        if (axisContainer) {
-            axisContainer.style.display = 'block';
-        }
-    } else if (mode === 'embedded') {
-        // Hide windowed version
-        const axisContainer = document.getElementById('axis-indicator-container');
-        if (axisContainer) {
-            axisContainer.style.display = 'none';
-        }
-        
-        // Check if we need to create the embedded indicator
-        if (!state.embeddedAxisIndicator) {
-            createEmbeddedAxisIndicator(scene, camera, renderer);
-        } else {
-            // Just reactivate it
-            state.embeddedAxisIndicator.active = true;
-        }
-        
-        // Update intensity if provided
-        if (state.embeddedAxisIndicator && intensity !== undefined) {
-            state.embeddedAxisIndicator.intensity = intensity;
-        }
-    } else if (mode === 'disabled') {
-        // Hide windowed version
-        const axisContainer = document.getElementById('axis-indicator-container');
-        if (axisContainer) {
-            axisContainer.style.display = 'none';
-        }
-        
-        // Deactivate embedded version
-        if (state.embeddedAxisIndicator) {
-            state.embeddedAxisIndicator.active = false;
-        }
-    }
-});
 
 // Global event listener for rig options changes
 document.addEventListener('rigOptionsChange', function(e) {
@@ -3301,16 +2535,7 @@ document.addEventListener('rigOptionsChange', function(e) {
         if (e.detail.hoverColor !== undefined) rigOptions.hoverColor = e.detail.hoverColor;
         if (e.detail.activeColor !== undefined) rigOptions.activeColor = e.detail.activeColor;
         
-        // Update control handle colors if it exists and not in a special state
-        if (furthestBoneHandle && furthestBoneHandle.material) {
-            // Only update color if it's not being hovered or dragged
-            if (furthestBoneHandle !== hoveredHandle && (!isDragging || furthestBoneHandle !== dragTarget)) {
-                furthestBoneHandle.material.color.setHex(rigOptions.normalColor);
-                furthestBoneHandle.material.needsUpdate = true;
-            }
-        }
-        
-        // Apply changes
+        // Apply the changes
         updateRigVisualization();
     }
 });
@@ -3320,3 +2545,16 @@ document.addEventListener('resetRig', function() {
     console.log('Reset rig event received');
     resetRig();
 });
+
+// Export functions needed by other modules
+export {
+    updateRigPanel,
+    updateRigAnimation,
+    updateAllBoneMatrices,
+    rigOptions,
+    createRigDetailsContent,
+    clearRigVisualization,
+    createAxisIndicator,
+    createJointLabels,
+    setupMouseListeners
+};
